adding existing sites to gropus now copies the site to the group site when appropriate

+++ objects/page.inc.php
@@ -57,8 +57,10 @@ class page extends segue {
 			
 			// remove stories
 			$this->fetchDown();
-			foreach ($this->stories as $s=>$o) {
-				$o->delete();
+			if ($this->stories) {
+				foreach ($this->stories as $s=>$o) {
+					$o->delete();
+				}
 			}
 			
 			$this->clearPermissions();
